refactor(notifications): extract kernel notifications to notification.rs
Move KernelNotificationInterfaceCallbacks, notification traits
(BlockTip, HeaderTip, Progress, WarningSet, WarningUnset, FlushError,
FatalError), and C wrapper fucntions to dedicated
notification.rs module.
