+++ src/notifications/mod.rs
@@ -1,0 +1,6 @@
+pub mod notification;
+
+pub use notification::{
+    BlockTip, FatalError, FlushError, HeaderTip, KernelNotificationInterfaceCallbacks, Progress,
+    WarningSet, WarningUnset,
+};
